requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
asyncpg>=0.29.0
openai>=1.35.3
pydantic>=2.7.4
python-dotenv>=1.0.1
apscheduler>=3.10.0
